#5335 - The application crashes when selecting Enhanced Stereochemistry from the right-click menu (#5342)
- added set stereo labels to atoms to setMolecule
